+++ src/message_sender.c
@@ -232,7 +232,7 @@ static SEND_ONE_MESSAGE_RESULT send_one_message(MESSAGE_SENDER_INSTANCE* message
                         }
                         else
                         {
-                            amqp_binary binary_value = { binary_data.bytes, binary_data.length };
+                            amqp_binary binary_value = { binary_data.bytes, (uint32_t)binary_data.length };
                             AMQP_VALUE body_amqp_data = amqpvalue_create_data(binary_value);
                             if (body_amqp_data == NULL)
                             {
@@ -321,7 +321,7 @@ static SEND_ONE_MESSAGE_RESULT send_one_message(MESSAGE_SENDER_INSTANCE* message
                         }
                         else
                         {
-                            amqp_binary binary_value = { binary_data.bytes, binary_data.length };
+                            amqp_binary binary_value = { binary_data.bytes, (uint32_t)binary_data.length };
                             AMQP_VALUE body_amqp_data = amqpvalue_create_data(binary_value);
                             if (body_amqp_data == NULL)
                             {
